+++ lib/sidekiq/client.rb
@@ -106,8 +106,8 @@ module Sidekiq
     # is run through the client middleware pipeline and each job gets its own Job ID
     # as normal.
     #
-    # Returns an array of the of pushed jobs' jids.  The number of jobs pushed can be less
-    # than the number given if the middleware stopped processing for one or more jobs.
+    # Returns an array of the of pushed jobs' jids, may contain nils if any client middleware
+    # prevented a job push.
     #
     # Example (pushing jobs in batches):
     #   push_bulk('class' => 'MyJob', 'args' => (1..100_000).to_a, batch_size: 10_000)
@@ -123,7 +123,6 @@ module Sidekiq
       raise ArgumentError, "Explicitly passing 'jid' when pushing more than one job is not supported" if jid && args.size > 1
 
       normed = normalize_item(items)
-
       result = args.each_slice(batch_size).flat_map do |slice|
         raise ArgumentError, "Bulk arguments must be an Array of Arrays: [[1], [2]]" unless slice.is_a?(Array) && slice.all?(Array)
         break [] if slice.empty? # no jobs to push
@@ -136,10 +135,11 @@ module Sidekiq
             copy
           end
           result || nil
-        }.compact
+        }
 
-        raw_push(payloads) unless payloads.empty?
-        payloads.collect { |payload| payload["jid"] }
+        to_push = payloads.compact
+        raw_push(to_push) unless to_push.empty?
+        payloads.map { |payload| payload&.[]("jid") }
       end
 
       result.is_a?(Enumerator::Lazy) ? result.force : result